Create unifiedServer.php
This server builds upon the addCredsServer.php and has a link to call the loginScript function. That function, loginScript, is held in a separate file as it needs its own .ini file to connect to the RabbitServer on the DB. After it runs and receives a response from the DB, the method returns either true or false, which is passed back to Apache2 and thrown up to the screen.